Show a dummy card before the ride starts
The ride controllers depend somewhat on the logger to exist before
the smart trainer can be controlled and ride started. It's
especially important for setting the right name for the activity
log. Because adding the logger as a dependency for the controller
is not nice we just render the controller once the logger is
initialized and this will initialize the actual controller code
too. This creates a layout shift in the dashboard which must
be filled with a place holder card for nicer UX.

Closes #157